==> DataStreamLine/01.convert_recording_to_video_with_timestamps.ipynb <==
import numpy as np
import pandas as pd
from pathlib import Path
import os
import matplotlib.pyplot as plt

import scipy.signal as signal
import imageio as io
from tqdm import tqdm
import cv2
font = cv2.FONT_HERSHEY_SIMPLEX
rgb = (255, 255, 255)

dirloc = '../../../../../../Documents/'
dir_converted_videos = 'converted_videos/'
fileloc_set = [
    '5_8_24/5_8_24_R72A10TrpA1Expt1_video_20240508_130433_636/',
    '5_8_24/5_8_24_R72A10TrpA1Expt2_video_20240508_142937_318/',
    '5_9_24/5_9_24_r72A10_expt1_video_20240509_095759_237/',
    '5_9_24/5_9_24_r72a10_expt2_5wellexpt_video_20240509_115237_346/',
    '5_13_24/5_13_24_r72A10_expt1_video_20240513_102826_071/',
    '5_13_24/5_13_24_r72A10_expt2_video_20240513_121809_465/',
    '5_13_24/5_13_24_r72A10_expt3_video_20240513_141847_637/',
    '5_15_24/5_15_24_r72a10_expt1_video_20240515_144359_162/',
    '5_17_24/5_17_24_r72a10_video_20240517_125705_729/',
#     '6_6_24/6_6_24_expt1_cooltemperature_video_20240606_114523_131/',
#     '6_11_24/6_11_24_cool_temp_video_20240611_113409_252/',
#     '6_12_24/6_12_24_Expt1_cool_temps_video_20240612_121128_422/',
#     '7_8_24/7_8_24_Expt1_video_20240708_101340_222/',
#     '7_8_24/7_8_24_Expt2_video_20240708_110603_123/',
#     '7_8_24/7_8_24_Expt3_video_20240708_130313_573/',
#     '7_9_24/7_9_24_expt1_video_20240709_114651_995/',
#     '7_10_24/7_10_24_expt1_video_20240710_124158_549/',
    '7_23_24/7_23_24_expt1_video_20240723_094225_800/',
    '7_23_24/7_23_24_expt2_video_20240723_130206_295/',
    '7_24_24/7_24_24_extp1_video_20240724_102123_758/',
    '7_24_24/7_24_24_extp2_video_20240724_120325_610/',
    '7_24_24/7_24_24_extp3_video_20240724_140346_190/',
    '7_24_24/7_24_24_extp4_video_20240724_144349_189/'
]

videofile_set = ['A1','B1','C1','D1',
                 'A2','B2','C2','D2']
exp_index_set = [
    '05_08_24_exp1',
    '05_08_24_exp2',
    '05_09_24_exp1',
    '05_09_24_exp2',
    '05_13_24_exp1',
    '05_13_24_exp2',
    '05_13_24_exp3', 
    '05_15_24_exp1',
    '05_17_24_exp1',
#     '06_06_24_exp1', 
#     '06_11_24_exp1',
#     '06_12_24_exp1',
#     '7_8_24_exp1',
#     '7_8_24_exp2',
#     '7_8_24_exp3',
#     '7_9_24_exp1',
#     '7_10_24_exp1',
    '07_23_24_exp1',
    '07_23_24_exp2',
    '07_24_24_exp1',
    '07_24_24_exp2',
    '07_24_24_exp3',
    '07_24_24_exp4',
]
print(len(fileloc_set), len(exp_index_set))
np.save('experiments_directory.npy',  fileloc_set)
np.save('experiments_well_setup.npy', videofile_set)
np.save('experiments_index.npy', exp_index_set)

# for fileloc in fileloc_set[17:23]:
#     if not os.path.exists(dirloc + dir_converted_videos + fileloc):
#         os.makedirs(dirloc + dir_converted_videos + fileloc) 
#     for videofile in videofile_set:
#         print(fileloc+'well_'+videofile+'.mp4')
#         vid = io.get_reader(dirloc+fileloc+'well_'+videofile+'.mp4')
#         height, width = vid.get_meta_data()['size']
#         frame_rate = int(vid.get_meta_data()['fps'])
#         frame_n    = vid.count_frames()
#         identifier = 'well_'+videofile+'_ts.mp4'
#         output_filepath = dirloc + dir_converted_videos + fileloc + identifier

#         writer = io.get_writer(output_filepath, fps = 60)
#         fn = 0
#         for frame in tqdm(vid.iter_data(), total=frame_n, desc=f"Analysing "+""):
#             foo = cv2.putText(frame, format(fn,'06d'), (5,520), font, .9, rgb)
#             writer.append_data(frame)
#             fn += 1
#         writer.close()
