==> DataStreamLine/02.a.data_entry_of_manual_score_edited.ipynb <==
import numpy as np
import pandas as pd
from pathlib import Path
import os
import matplotlib.pyplot as plt

import scipy.signal as signal
import imageio as io
from tqdm import tqdm
import cv2


dirloc = '../../../../../../Documents/'
fileloc_set   = np.load('experiments_directory.npy')
videofile_set = np.load('experiments_well_setup.npy')
experiments_index = np.load('experiments_index.npy')
print(fileloc_set, len(fileloc_set))
print("")
print(videofile_set)
print("")
print(experiments_index)


experiment_mscore= {}
experiments_mscore= {}
# exclusion -1
# female 0, male 1, 
# R72xTrpA1 0, R72xW 1, Trp1AxW 2
# HB 1, UWE 2, Lunge 3, Non, 0
# floor 0, wall 1, glass 2

# 05-08-2024-exp1

experiment_info = {}
event_set = []
event_set_Type = []
event_set_Loc = []

event = np.array([
    3991,5314,13908,14643,14839,14844,19005,19467,26253,30251,36777,51032,64188,65123,
    65133,65223,65789,73196,90892,92389,96628
])
event_set.append(event) #A1 female
event_Type = np.ones(event.shape)
event_Type[[0,1,3,5,8,9,10]] = 0
event_set_Type.append(event_Type)
event_Loc = np.ones(event.shape)*1
event_Loc[0:12] = 0
event_set_Loc.append(event_Loc)


event = np.array([
    63216,63674,64059,
])
event_set.append(event) #B1 male
event_Type = np.ones(event.shape)*2
event_set_Type.append(event_Type)
event_Loc = np.ones(event.shape)*0
event_set_Loc.append(event_Loc)



event = np.array([])
event_set.append(event) #C1
event_Type = np.array([])
event_set_Type.append(event_Type)
event_Loc = np.array([])
event_set_Loc.append(event_Loc)


event = np.array([11201,14504,14620,15016,17071,17408,17799,24128,25400,53311,63434,67847,82940,83838,91932,96718,121286])
event_set.append(event) #D1 male
event_Type = np.ones(event.shape)*2
event_Type[[0,3]] = 0
event_set_Type.append(event_Type)

event_Loc = np.ones(event.shape)*0
event_Loc[[0,8]] = 1
event_Loc[9] = 2
event_set_Loc.append(event_Loc)

event = np.array([7721,10406,15989,31368,71583])
event_set.append(event) #A2 female

event_Type = np.ones(event.shape)
event_Type[[0,3]] = 0
event_set_Type.append(event_Type)
event_Loc = np.ones(event.shape)*1
event_Loc[[2,3]] = 0
event_set_Loc.append(event_Loc)


event = np.array([16721,17141,17998,49252,50381,56023,61590,63424,93808,100400])
event_set.append(event) #B2 female

event_Type = np.ones(event.shape)
event_Type[[3,4]] = 0
event_set_Type.append(event_Type)
event_Loc = np.ones(event.shape)
event_Loc[5] = 0
event_set_Loc.append(event_Loc)


event = np.array([])
event_set.append(event) #C2
event_Type = np.array([])
event_set_Type.append(event_Type)
event_Loc = np.array([])
event_set_Loc.append(event_Loc)

event = np.array([])
event_set.append(event) #D2
event_Type = np.array([])
event_set_Type.append(event_Type)
event_Loc = np.array([])
event_set_Loc.append(event_Loc)

exp_id = 0
experiment_mscore[fileloc_set[exp_id]] = event_set

experiment_info['events'] = event_set
experiment_info['events type'] = event_set_Type
experiment_info['events loc']  = event_set_Loc
experiment_info['sex']      = [0,1,-1,1,0,0,-1,-1]
experiment_info['genotype'] = [2,1,-1,0,2,1,-1,-1]
experiment_info['starting_time'] = [2311,3569,4246,5148,6409,6214,5943,5519]
experiments_mscore[fileloc_set[exp_id]] = experiment_info

# 05-08-2024-exp2

experiment_info = {}
event_set = []
event_set_Type = []
event_set_Loc = []

event = np.array([])
event_set.append(event) #A1
event_Type = np.array([])
event_set_Type.append(event_Type)
event_Loc = np.array([])
event_set_Loc.append(event_Loc)

event = np.array([])
event_set.append(event) #B1
event_Type = np.array([])
event_set_Type.append(event_Type)
event_Loc = np.array([])
event_set_Loc.append(event_Loc)

event = np.array([
    18307,19343,21552,24745,27406,27465,28020,30234,31538,33347,
33928,34368,34618,36029,36224,36449,36466,36569,37460,40036,40733,42340,49364,51603,51740,52334,52610,53251,
54270,55621,56517,56792,59447,60395,60445,61070,61128,61739,62719,
    63580,63623,65659,66707,67465,67702,67920,75627,77705,80667,82434,83503,84718,86306,87114,87666,93402,93703,
94780,95130,96041,96046,98233,100728,101077,103230,105167,107873,107978,113152,114124,114149,116398,116743
])
event_set.append(event) #C1 female
event_Type = np.ones(event.shape)
event_Type[[1,7,10,15,19,20,27,29,33,34,40,46,57,59,71]] = 0
event_set_Type.append(event_Type)
event_Loc = np.ones(event.shape)*0
event_set_Loc.append(event_Loc)

event = np.array([
17382,17425,17551,17564,24449,24469,26298,
26342,40694,41266,41309,41495,44539,44630,49460,54706,55088,56966,69337,69539,
75393,77315,77339,85309,85333,85762,88955,89483,94445,94515,95125,95163,101694,
101709,102293,102320,105377,106979,110482,110504,110562,115704,115920,117796
])
event_set.append(event) #D1 female
event_Type = np.ones(event.shape)
event_Type[[14,26,38,41]] = 0
event_set_Type.append(event_Type)
event_Loc = np.ones(event.shape)*0
event_Loc[26] = 1
event_set_Loc.append(event_Loc)

event = np.array([
5756,13846,14830,24049,30931,31723,34174,34196,34302,34388,34759,34966,36368,36372,36512,36528,
36696,36725,36981,38639,38707,39359,39372,39481,39870,40092,40644,40702,40809,41757,41800,41821,
42017,56924,58131,58142,58389,58608,60006,60015,60060,60403,62883,63066,64221,64380,68716,73582,
74636,85899,86556,97476,97520,97781,97969,105371,105427,109344,109357,109785,110305,110465,111874,111969,
112054,112834,114324,114802,115905,115938,115987,116067,117890,118519,119857
])
event_set.append(event) #A2 male
event_Type = np.ones(event.shape)*3
event_Type[[1,2]] = 2
event_Type[[0,4,13,25,32,33,45,66]] = 0
event_set_Type.append(event_Type)
event_Loc = np.ones(event.shape)*0
event_Loc[[1,2]] = 2
event_set_Loc.append(event_Loc)



event = np.array([20994,22359])
event_set.append(event) #B2 male
event_Type = np.zeros(event.shape)
event_set_Type.append(event_Type)
event_Loc = np.ones(event.shape)
event_set_Loc.append(event_Loc)

event = np.array([])
event_set.append(event) #C2
event_Type = np.array([])
event_set_Type.append(event_Type)
event_Loc = np.array([])
event_set_Loc.append(event_Loc)

event = np.array([
19529,27688,29738,29901,35619,39089,39103,39128,39814,40370,40395,40586,41550,45911,47765,62809,
63796,64473,64480,64552,71263,72885,78470,78515,79988,80031,80035,80131,80140,80151,81408,81430,
81437,81906,81909,81930,81966,82087,92070,95558,96543,96600,96737,99325,101282,101295,102453,102469,
109659,109722,110045,110077,110170,111218,111309,111423,111427,111435,111447,111459,111994,112797,113558,113896,
113995,115218,115239,115267,115294,115321,115370,115941,116465
])
event_set.append(event) #D2 female
event_Type = np.ones(event.shape)
event_Type[[0,4,10,13,38,52,64,65]] = 0
event_set_Type.append(event_Type)
event_Loc = np.ones(event.shape)*0
event_Loc[[1,2,3,14,15,20,21]] = 1
event_Loc[38:52] = 1
event_set_Loc.append(event_Loc)

exp_id = 1

experiment_mscore[fileloc_set[exp_id]] = event_set
experiment_info['events'] = event_set
experiment_info['events type'] = event_set_Type
experiment_info['events loc']  = event_set_Loc
experiment_info['sex']         = [-1,-1,0,0,1,1,-1,0]
experiment_info['genotype']    = [-1,-1,0,0,2,1,-1,0]
experiment_info['starting_time'] = [1913,2168,2637,3095,4151,4399,4822,5088]
experiments_mscore[fileloc_set[exp_id]] = experiment_info

# 05-09-2024-exp1

experiment_info = {}
event_set = []
event_set_Type = []
event_set_Loc = []

event = np.array([
4685,6432,15132,16728,33082,34261,34554,35163,35654,35719,37278,38292,38817,41305,46997,84925,
85037,87383,101362,101590,115014,115107,115244,115545,115950,117292,117622,117666
])
event_set.append(event) #A1 male
event_Type = np.ones(event.shape)*2
event_Type[[0,1]] = 0
event_Type[[13]] = 3
event_set_Type.append(event_Type)

event_Loc = np.ones(event.shape)*0
event_Loc[[0,1,6,7,10,11,12,13,24,25]] = 1
event_set_Loc.append(event_Loc)


event = np.array([])
event_set.append(event) #B1

event_Type = np.array([])
event_set_Type.append(event_Type)
event_Loc = np.array([])
event_set_Loc.append(event_Loc)


event = np.array([
9140,10600,32467,32495,32525,33168,33303,33336,33465,33846,33990,35477,36584,36784,38988,39689,
40071,41448,41490,42179,42542,44493,46720,49126,49727,51711,52104,52604,52667,52688,52709,53317,
56884,60018,60045,76691,76710,76719,76942,77168,77631,81937,123232
])
event_set.append(event) #C1 female
event_Type = np.ones(event.shape)
event_Type[[4,21,28]] = 0
event_set_Type.append(event_Type)
event_Loc = np.ones(event.shape)*0
event_Loc[[0]] = 1
event_set_Loc.append(event_Loc)

event = np.array([21002,22473,28352,97312])
event_set.append(event) #D1 male
event_Type = np.zeros(event.shape)
event_set_Type.append(event_Type)
event_Loc = np.ones(event.shape)*0
event_set_Loc.append(event_Loc)

event = np.array([22294,23626,24028,24349,24707,26715, 27037,28643,36629,65069,68526,68757,74095,106048,
                  107812,108723,109404,111801,112677,112903])
event_set.append(event) #A2 male
event_Type = np.ones(event.shape)*2
event_Type[[6,10]] = 0
event_set_Type.append(event_Type)
event_Loc = np.ones(event.shape)*0
event_Loc[[0,17]] = 1
event_set_Loc.append(event_Loc)


event = np.array([
5096,13241,14540,14560,14588,18606,27848,29719,29766,31412,33746,33763,41690,41853,44265,44579,
44833,45350,55566,55584,73835,81368,82084,83548,83628,111181,121912
])#67190 wall headbutt
event_set.append(event) #B2 female
event_Type = np.ones(event.shape)
event_Type[[4]] = 0
event_set_Type.append(event_Type)
event_Loc = np.ones(event.shape)*0
event_Loc[[15,16,17,20,21,22,23]] = 1
event_set_Loc.append(event_Loc)



event = np.array([])
event_set.append(event) #C2 male
event_Type = np.array([])
event_set_Type.append(event_Type)
event_Loc = np.array([])
event_set_Loc.append(event_Loc)

event = np.array([30910,31496,74709,107618])
event_set.append(event) #D2 female
event_Type = np.ones(event.shape)
event_set_Type.append(event_Type)
event_Loc = np.ones(event.shape)*0
event_set_Loc.append(event_Loc)

exp_id = 2
experiment_mscore[fileloc_set[exp_id]] = event_set


experiment_info['events'] = event_set
experiment_info['events type'] = event_set_Type
experiment_info['events loc']  = event_set_Loc
experiment_info['sex']      = [1,-1,0,1,1,0,1,0]
experiment_info['genotype'] = [0,-1,2,1,0,0,2,1]
experiment_info['starting_time'] = [2521, 2857, 3072, 3330,4032,4496,4725,4885]
experiments_mscore[fileloc_set[exp_id]] = experiment_info

# 05-09-2024-exp2

experiment_info = {}
event_set = []
event_set_Type = []
event_set_Loc = []

event = np.array([])
event_set.append(event) #A1
event_Type = np.array([])
event_set_Type.append(event_Type)
event_Loc = np.array([])
event_set_Loc.append(event_Loc)

event = np.array([47026,47100,47162,47527,75282,76903,116159,116430])
event_set.append(event) #B1 female
event_Type = np.ones(event.shape)
event_set_Type.append(event_Type)
event_Loc = np.ones(event.shape)*1
event_Loc[[3,6,7]] = 0
event_set_Loc.append(event_Loc)

event = np.array([10851])
event_Type = np.zeros(event.shape)
event_set_Type.append(event_Type)
event_set.append(event) #C1 male
event_Loc = np.ones(event.shape)*0
event_set_Loc.append(event_Loc)

event = np.array([
    65236,65889,70005,70315,70397,72503,72648,73073,73096,73216,74370,74399,90672,90771,90828,91102,92195,92254,
92275,95452,95880,96545,96593,97409,99441,102811,104834,113397,120085,120133,123216,123194
])
event_set.append(event) #D1 male
event_Type = np.ones(event.shape)*3
event_set_Type.append(event_Type)
event_Loc = np.ones(event.shape)*0
event_Loc[[19,27]] = 1
event_set_Loc.append(event_Loc)

event = np.array([6950,7027,7300,8399,8578,9198,10596,11171,12887,12999,13670,13730,23561,23728,24083,26398,
27198,27795,29242,29286,31994,32208,34054,34817,39833,39865,39941,44060,50193,52822,56058,56095,
56236,56258,56274,56604,64317,69443,96632,116206,116243])
event_set.append(event) #A2 female
event_Type = np.ones(event.shape)*1
event_Type[[28,38]] = 0
event_set_Type.append(event_Type)
event_Loc = np.ones(event.shape)*0
event_Loc[[5,37,39,40]] = 1
event_set_Loc.append(event_Loc)

event = np.array([])
event_set.append(event) #B2
event_Type = np.array([])
event_set_Type.append(event_Type)
event_Loc = np.array([])
event_set_Loc.append(event_Loc)

event = np.array([])
event_set.append(event) #C2
event_Type = np.array([])
event_set_Type.append(event_Type)
event_Loc = np.array([])
event_set_Loc.append(event_Loc)

event = np.array([])
event_set.append(event) #D2
event_Type = np.array([])
event_set_Type.append(event_Type)
event_Loc = np.array([])
event_set_Loc.append(event_Loc)


exp_id = 3
experiment_mscore[fileloc_set[exp_id]] = event_set

experiment_info['events'] = event_set
experiment_info['events type'] = event_set_Type
experiment_info['events loc']  = event_set_Loc
experiment_info['sex']      = [-1,0,1,1,0,-1,-1,-1]
experiment_info['genotype'] = [-1,1,2,2,1,-1,-1,-1]
experiment_info['starting_time'] = [3069,3296,3505,3754,2000,2240,2448,2585]
experiments_mscore[fileloc_set[exp_id]] = experiment_info

# 05-13-2024-exp1

experiment_info = {}
event_set = []
event_set_Type = []
event_set_Loc = []

event = np.array([10167,10190,10451,10465,10510,11938,13511,13929,14900,15315,17842,17974,18715,
19942,20417,21448,22322,28016,32582,32912,34126,38136,38603,41649,42296,43134,43864,44542,48721,
48947,49205,50069,50477,51106,53361,53586,53852,53995,61712,61847,62008,63140,63427,63761,64817,
65160,67114,68061,68139,68348,68663,68775,69158,69194,69444,69577,70847,74393,75991,79260,79662,
80019,80252,81995,82961,83479,84110,87358,87594,87700,88473,88694,88991,90815,91387,92773,93593,
94037,95099,95426,95699,95904,96452,96991,98653,98845,99801,100125,100315,100789,101019,101214,101589,
104574,106764,108443,109050,109518,111325,113114,115810,
117869,118236,118461,123219,125207,125356,125599])
event_set.append(event) #A1

event_Type = np.ones(event.shape)*2
event_Type[[1,2,3,4]] = 3
event_Type[[5,11,12,35,45,46,51,52,64,83,97,99,107]] = 0
event_set_Type.append(event_Type)
event_Loc = np.ones(event.shape)*1
event_Loc[[5,13,23,24,27,32,33,35,37,38,39,50,51,52,53,57,58,64,70,73,75,83,93,96,100]] = 0
event_Loc[[22,26,45,60]] = 2
event_set_Loc.append(event_Loc)

event = np.array([])
event_set.append(event) #B1
event_Type = np.array([])
event_set_Type.append(event_Type)
event_Loc = np.array([])
event_set_Loc.append(event_Loc)

event = np.array([])
event_set.append(event) #C1
event_Type = np.array([])
event_set_Type.append(event_Type)
event_Loc = np.array([])
event_set_Loc.append(event_Loc)

event = np.array([
10646,11010,11544,14420,14935,17178,17881,28268,28907,29424,29908,31330,
31859,34348,38470,38840,40127,40500,41020,55248,72808,78705,78996,84330,
84808,85345,85896,86618,86854,87332,88475,89233,89726,93506,93838,94301,
94662,95270,95638,97008,97737,103639,103673,105922,106512,109820,110229,112513,
114117,114162,114188,115426,116863,118904,119235,119482,119927,120094
])
event_set.append(event) #D1
event_Type = np.ones(event.shape)*2
event_Type[[41,42,48,49,50]] = 3
event_Type[[5,6,21]] = 0
event_set_Type.append(event_Type)
event_Loc = np.ones(event.shape)*0
event_set_Loc.append(event_Loc)

event = np.array([])
event_set.append(event) #A2
event_Type = np.array([])
event_set_Type.append(event_Type)
event_Loc = np.array([])
event_set_Loc.append(event_Loc)

event = np.array([33254,55434,55566,61637,61708,61730,62515,83242])
event_set.append(event) #B2
event_Type = np.ones(event.shape)
event_set_Type.append(event_Type)
event_Loc = np.ones(event.shape)*0
event_Loc[[7]] = 1
event_set_Loc.append(event_Loc)

event = np.array([])
event_set.append(event) #C2
event_Type = np.array([])
event_set_Type.append(event_Type)
event_Loc = np.array([])
event_set_Loc.append(event_Loc)

event = np.array([31951,33959,34330,34700,40385,42800,44503,44559,44739,45503,46560,47901,48065,48648,51817,67639,
74159,74717,75119,75140,76152,88441,88474,96041,103442,104135,105060,105186,117377])
event_set.append(event) #D2
event_Type = np.ones(event.shape)
event_Type[[22]] = 3
event_Type[[1,7]] = 0
event_set_Type.append(event_Type)
event_Loc = np.ones(event.shape)*0
event_Loc[[16]] = 1
event_set_Loc.append(event_Loc)

exp_id = 4
experiment_mscore[fileloc_set[exp_id]] = event_set


experiment_info['events'] = event_set
experiment_info['events type'] = event_set_Type
experiment_info['events loc']  = event_set_Loc
experiment_info['sex'] = [1,-1,-1,1,1,0,-1,0]
experiment_info['genotype'] = [0,-1,-1,1,2,1,-1,0]
experiment_info['starting_time'] = [3623, 3813,4502, 4868,2236,2491,0,2899]
experiments_mscore[fileloc_set[exp_id]] = experiment_info

# 05-13-2024-exp2

experiment_info = {}
event_set = []
event_set_Type = []
event_set_Loc = []

event = np.array([])
event_set.append(event) #A1
event_Type = np.array([])
event_set_Type.append(event_Type)
event_Loc = np.array([])
event_set_Loc.append(event_Loc)

event = np.array([8856,8866,8905,9250,10187,10359,13099,13424,14561,16474,16690,17127,17203,17523,18876,22844,
24811,26304,27209,28635,29710,37795,38134,38717,40047,41925,42908,43865,46065,56313,56684,57375,
57573,57613,58144,60748,60808,61111,61214,65085,65752,65892,67945,84192,86265,89668,98062,104757,
105481,106061,108358,109390,110779,119068,119071,119119])
event_set.append(event) #B1
event_Type = np.ones(event.shape)
event_Type[[5,10,12,31,32,35,55]] = 0
event_set_Type.append(event_Type)
event_Loc = np.ones(event.shape)*0
event_set_Loc.append(event_Loc)

event = np.array([41015,41024,41048,41060,45015,46270,46382,46460,48266,48327,48389,48484,52873,52881,52907,54399,
56265,56830,61847,67084,67113,68000,72931,78786,90316,90680,91495,91510,91566,92993,95360,95383,
95456,97436,97448,97512,97532,99377,99402,100897,100942,101154,102486,102524,105122,106887,111310,112017,
112125,112154,113265,113548,113663,114437,114458,114482,114520,114987,115805,116165,116328,116696,116756,117082,
117559,120569,121030,121766,122137,122327,122621,123950,124396,124668,125052,125460])
event_set.append(event) #C1
event_Type = np.ones(event.shape)*3
event_Type[[35]] = 0
event_Type[[4,23,30,33,47,67,68,69,70,71,72,73,74,75]] = 2
event_set_Type.append(event_Type)
event_Loc = np.ones(event.shape)*0
event_set_Loc.append(event_Loc)

event = np.array([])
event_set.append(event) #D1
event_Type = np.array([])
event_set_Type.append(event_Type)
event_Loc = np.array([])
event_set_Loc.append(event_Loc)

event = np.array([])
event_set.append(event) #A2
event_Type = np.array([])
event_set_Type.append(event_Type)
event_Loc = np.array([])
event_set_Loc.append(event_Loc)

event = np.array([])
event_set.append(event) #B2
event_Type = np.array([])
event_set_Type.append(event_Type)
event_Loc = np.array([])
event_set_Loc.append(event_Loc)

event = np.array([8792,9522,12665,12892,12926,13696,15235,17013,17327,18231,18312,18475,18991,19536,19793,20410,
20810,20986,21840,22174,22617,22810,22949,23435,23582,23829,24645,24719,24877,25300,26026,27719,
27880,28175,28349,28377,28621,28956,29341,29400,30495,30689,30982,31223,31487,32764,34229,35917,
41454,41985,43881,44462,47429,47715,47976,48200,48497,49047,49288,49329,50673,51228,51688,53539,
53884,54336,54349,54591,55729,56325,56709,59284,59568,59829,61154,61789,62752,64745,65634,65651,
67741,70190,71800,71982,72069,75042,78874,79701,79923,81094,82532,84329,84551,84830,85051,85301,
86335,86678,86773,87272,87512,87592,87710,87913,88323,100236,105197,105434,106950,107580,107860,108082,
108290,109505,110585,112726,113003,117579,119605])
event_set.append(event) #C2
event_Type = np.ones(event.shape)*2
event_Type[[15,33,35,90,101]] = 0
event_Type[[3,4,5,10,11,12,13,50,51]] = 3
event_set_Type.append(event_Type)
event_Loc = np.ones(event.shape)
event_Loc[[1,62,63,77,99,114]] = 2
event_Loc[[24,28,30,31,33,35,65,66,68,75,78,79,81,82,83,89,90,96,97,100,107]] = 0
event_set_Loc.append(event_Loc)

event = np.array([])
event_set.append(event) #D2
event_Type = np.array([])
event_set_Type.append(event_Type)
event_Loc = np.array([])
event_set_Loc.append(event_Loc)


exp_id = 5
experiment_mscore[fileloc_set[exp_id]] = event_set

experiment_info['events'] = event_set
experiment_info['events type'] = event_set_Type
experiment_info['events loc']  = event_set_Loc
experiment_info['sex'] = [-1,0,1,1,-1,-1,1,-1]
experiment_info['genotype'] = [-1,2,1,1,-1,-1,0,-1]
experiment_info['starting_time'] = [0,3485,3671,3862,2192,2675,2795,0]
experiments_mscore[fileloc_set[exp_id]] = experiment_info

# 05-13-2024-exp3

experiment_info = {}
event_set = []
event_set_Type = []
event_set_Loc = []

event = np.array([
    4345,4900,6751,7683,11620,11794,13203,15034,15785,17090,17705,17944,17971,18886,19183,19764,
19966,21064,22096,22591,22817,23315,23811,24931,24992,27031,27217,27423,28589,29886,30608,30982,
31823,32562,33593,33924,34930,35354,35569,35927,36418,36957,37234,38975,41411,41910,42976,43262,
43916,48453,49700,53218,56642,57374,58934,59249,59426,59780,62089,62466,62932,64489,66017,66132,
66386,67068,68443,69392,73552,73647,73901,79697,81610,89464,92567,94375,105113,116437,117317,120027,
121635])
event_set.append(event) #A1
event_Type = np.ones(event.shape)*2
event_Type[[12,18,61]] = 0
event_Type[[2]] = 3
event_set_Type.append(event_Type)
event_Loc = np.ones(event.shape)
event_Loc[[0,2,3,12,18,19,20,28,32,33,37,41,47,49,54,56,57,58,60,
          61,66,67,69,70,73,74,77,78,80]] = 0
event_set_Loc.append(event_Loc)

event = np.array([
    5554,5683,5983,6064,6087,6228,7018,7188,7326,7501,7852,7918,8484,8907,10311,12634,
12658,13531,14143,14167,14189,15245,16154,19883,19932,20063,20552,20973,23976,24060,26339,26372,
26384,26607,27865,30535,30534,34628,34794,41534,43253,44838,45485,47826,48825,54631,56601,57848,
59995,60018,64065,68596,68610,68625,68676,72156,72833,75871,75898,75924,76473,76508,76936,77621,
79689,80370,82687,84136,84165,85564,90107,91776,93353,95371,96518,96793,97945,99557,104466,104501,
110969,111623,115950,116425,123550
])
event_set.append(event) #B1
event_Type = np.ones(event.shape)*1
event_Type[[3,4,11,12,13,56,60,61,64]] = 0
event_Type[[64]] = 3
event_set_Type.append(event_Type)
event_Loc = np.ones(event.shape)*0
event_Loc[[81]] = 1
event_set_Loc.append(event_Loc)

event = np.array([
    32591,51758,52941,66513,71319,95175
])
event_set.append(event) #C1
event_Type = np.ones(event.shape)*2
event_Type[[3]] = 0
event_Type[[2,4]] = 3
event_set_Type.append(event_Type)
event_Loc = np.ones(event.shape)*0
event_Loc[[0]] = 1
event_set_Loc.append(event_Loc)

event = np.array([
    14576,15029,17093,20522,23356,30764,32562,32766,34372,34702,45516,45844,48408,49582,49999,51495,
52956,53323,53445,53856,54411,58697,59304,66912,67334,67819,68213,71641,73327,73993,76232,82986,
86909,89685,90331,93331,94199,95404,96169,99789,103477,105434,109684,116542,117325,119243,120986,121294,
121971,123900,124085,124315
])
event_set.append(event) #D1
event_Type = np.ones(event.shape)*2
event_Type[[17]] = 0
event_set_Type.append(event_Type)
event_Loc = np.ones(event.shape)*1
event_Loc[[0,1,4,8,19]] = 2
event_Loc[[5,6,7,9,15,16,17,23,24,25,26,27,29,30,39]] = 0
event_set_Loc.append(event_Loc)

event = np.array([])
event_set.append(event) #A2
event_Type = np.array([])
event_set_Type.append(event_Type)
event_Loc = np.array([])
event_set_Loc.append(event_Loc)

event = np.array([])
event_set.append(event) #B2
event_Type = np.array([])
event_set_Type.append(event_Type)
event_Loc = np.array([])
event_set_Loc.append(event_Loc)

event = np.array([])
event_set.append(event) #C2
event_Type = np.array([])
event_set_Type.append(event_Type)
event_Loc = np.array([])
event_set_Loc.append(event_Loc)

event = np.array([])
event_set.append(event) #D2
event_Type = np.array([])
event_set_Type.append(event_Type)
event_Loc = np.array([])
event_set_Loc.append(event_Loc)


exp_id = 6
experiment_mscore[fileloc_set[exp_id]] = event_set

experiment_info['events'] = event_set
experiment_info['events type'] = event_set_Type
experiment_info['events loc']  = event_set_Loc
experiment_info['sex'] = [1,0,1,1,-1,-1,-1,-1]
experiment_info['genotype'] = [0,2,1,0,-1,-1,-1,-1]
experiment_info['starting_time'] = [1748,2138,2350,3142,0,0,0,0]
experiments_mscore[fileloc_set[exp_id]] = experiment_info

# 05-15-2024-exp1

experiment_info = {}
event_set = []
event_set_Type = []
event_set_Loc = []

event = np.array([
    33865,51227,54427,64166,65345,76951
])
event_set.append(event) #A1 male
event_Type = np.ones(event.shape)*3
event_set_Type.append(event_Type)
event_Loc      = np.ones(event.shape)*0
event_set_Loc.append(event_Loc)


event = np.array([
    62661,65273,104015,115227
])
event_set.append(event) #B1
event_Type = np.ones(event.shape)*3
event_Type[[0,1]] = 0
event_set_Type.append(event_Type)
event_Loc = np.ones(event.shape)*0
event_Loc[[0,1]] = 1
event_set_Loc.append(event_Loc)

event = np.array([
    56136,107873,109202,109277,110356,115993,116076
])
event_set.append(event) #C1
event_Type = np.ones(event.shape)*3
event_set_Type.append(event_Type)
event_Loc = np.ones(event.shape)*0
event_set_Loc.append(event_Loc)

event = np.array([
    26206,81009,83208,91142,103224,111684
])
event_set.append(event) #D1
event_Type = np.ones(event.shape)*1
event_set_Type.append(event_Type)
event_Loc = np.ones(event.shape)*0
event_Loc[[0,5]] = 1
event_set_Loc.append(event_Loc)


event = np.array([
    51354,96331,104796,117545,120851,120883
])
event_set.append(event) #A2
event_Type = np.ones(event.shape)*3
event_Type[[0,2]] = 0
event_set_Type.append(event_Type)
event_Loc = np.ones(event.shape)*0
event_Loc[[0]] = 1
event_set_Loc.append(event_Loc)

event = np.array([
    55554,62251,66011,80834,81373,82024,82109,89374,89422,90903,92789,92818,92830,95984,96093,96210,
    96218,96240,97825,97855,97891,97973,97998,98037,98073,121121,121191,121233,121278,122285,123054,123769
])
event_set.append(event) #B2
event_Type = np.ones(event.shape)*3
event_Type[[2,31]] = 2
event_Type[[1]] = 0
event_set_Type.append(event_Type)
event_Loc = np.ones(event.shape)*0
event_set_Loc.append(event_Loc)

event = np.array([
    31939,34728,34741
])
event_set.append(event) #C2
event_Type = np.ones(event.shape)*1
event_set_Type.append(event_Type)
event_Loc = np.ones(event.shape)*0
event_set_Loc.append(event_Loc)

event = np.array([
    22965,22986,23206,26504,26525,26526,30391,30816,36463,37194,40607,43408,49398,49914,49937,53129,53308,
    53334,70818,70832,70865,86923,92863,97527,101256,113326,120207,120254,120283,120622,120969,121649,123421
])
event_set.append(event) #D2
event_Type = np.ones(event.shape)*1
event_set_Type.append(event_Type)
event_Loc = np.ones(event.shape)*0
event_set_Loc.append(event_Loc)


exp_id = 7
experiment_mscore[fileloc_set[exp_id]] = event_set
experiment_info['events'] = event_set
experiment_info['events type'] = event_set_Type
experiment_info['events loc']  = event_set_Loc
experiment_info['sex']      = [1,1,1,0,1,1,0,0]
experiment_info['genotype'] = [2,2,2,0,1,1,0,0]
experiment_info['starting_time'] = [13794,14095,14265,14368,14921,15120,15460,15649]
experiments_mscore[fileloc_set[exp_id]] = experiment_info

# 05-17-2024-exp1

experiment_info = {}
event_set = []
event_set_Type = []
event_set_Loc = []

event = np.array([
    6383,14942,14991,15151,15825,15880,16630,16639,16660,16981,17033,17751,17789,18311,19600,20121,20350,20440,
    21553,21625,26278,27728,27783,28642,31164,31672,31687,31710,32293,32353,36666,36739,36806,38804,39383,
    39602,39991,44451,44465,44504,44587,45661,46738,46824,46867,48581,48602,49375,52967,53039,53053,53892,53969,
    60364,60401,60433,60463,61137,61173,62086,62757,68010,68057,68067,71243,72437,72972,73329,75108,
    75492,75859,75955,76211,77275,77458,78252,79035,79074,79228,79569,79650,79945,81909,82852,82882,83544,
    83657,83707,83725,84885,85938,88566,90208,91466,92877,92973,95756,96351,96474,98317,103764,104214,
    110491,112138,112154,112687,117041,117240,12348,123521
])
event_set.append(event) #A1 female
event_Type = np.ones(event.shape)*1
event_Type[[9,13,44,51,52,56,60,66,68,69,78,90,91,99,105]] = 0
event_set_Type.append(event_Type)
event_Loc  = np.ones(event.shape)*0
event_Loc[[75,104,105]] = 1
event_set_Loc.append(event_Loc)

event = np.array([
    2380,2761,3377,12600,15372,15595,15787,15926,16224,17028,17082,17110,17121,17174,23477,27035,28060,
    29031,32755,32848,33867,34273,34370,35734,35795,40261,43559,43594,48936,49774,53678,59860,
    61903,62577,64605,70233,71669,74126,76943,90882,91725,93454,98314,99555,105246,110716,111897,122578
])
event_set.append(event) #B1 male
event_Type = np.ones(event.shape)*3
event_Type[[3,43]] = 2
event_Type[[0,1,2,4,5,15,20,25,28]] = 0
event_set_Type.append(event_Type)
event_Loc  = np.ones(event.shape)*0
event_Loc[[8,32,41]] = 1
event_Loc[[3,43]] = 2
event_set_Loc.append(event_Loc)


event = np.array([
28138,45862,47723,47783,48527,54723,60273,61934,62123,64809,64897,67817,70412,76711,77634,80695,
81039,85060,85887,86038,88281,92859,98178,101711,103618,104983,105040,111880,115261,116055,118009,118090,
121252,122452,124201,124906
])
event_set.append(event) #C1 male
event_Type = np.ones(event.shape)*3
event_set_Type.append(event_Type)
event_Loc  = np.ones(event.shape)*0
event_Loc[[0,21]] = 1
event_set_Loc.append(event_Loc)

event = np.array([
20867,23002,23144,23689,26135,32107,32159,32474,33015,38107,39910,40111,44480,46145,48307,48397,
51382,51442,54926,55071,55728,55737,56457,57271,57773,59939,59987,60911,61730,61749,63497,64572,
64590,64718,65346,67214,68033,68380,68730,69161,70968,72888,73051,77418,78453,80371,81895,82359,
82584,82589,82881,83156,84733,85262,85314,86235,87061,87580,87979,88023,89638,91094,91467,91763,
92826,97918,98233,98287,102705,103279,104761,104980,105602,110452,110497,110830,111247,112803,112822,113083,
114473,116311,116657,116702,118465,120708,121845,122265,122286,123904,123728,125315
])
event_set.append(event) #D1
event_Type = np.ones(event.shape)*3
event_Type[[3]] = 2
event_Type[[67,70,71,88]] = 0
event_set_Type.append(event_Type)
event_Loc  = np.ones(event.shape)*0
event_Loc[[81]] = 1
event_set_Loc.append(event_Loc)

event = np.array([
6542,7155,8215,9417,18232,19436,23601,25046,33467,36891,52579,52616,80812,85092,89503,97180,110532
])
event_set.append(event) #A2 female
event_Type = np.ones(event.shape)*1
event_Type[[0,3,12,13,14]] = 0
event_set_Type.append(event_Type)
event_Loc  = np.ones(event.shape)*0
event_Loc[[7]] = 1
event_set_Loc.append(event_Loc)

event = np.array([
15635,16647,17423,19771,20230,21437,23882,26726,29514,34373,39698,39719,40331,40870,41738,42623,
50454,51559,51601,56429,58981,63548,68300,68875,73158,76473,76664,79584,80812,81222,84568,87241,
96098,101226,108587,118581,120635,120736,125425
])
event_set.append(event) #B2 female
event_Type = np.ones(event.shape)*1
event_Type[[8,22]] = 0
event_set_Type.append(event_Type)
event_Loc  = np.ones(event.shape)*0
event_Loc[[15,35]] = 1
event_set_Loc.append(event_Loc)

event = np.array([
12962,22529,28156,28221,28662,86246,86316,86404,86792,87377,92652,98166,114106,119914,120022,120275,
121872,122130,122234
])
event_set.append(event) #C2
event_Type = np.ones(event.shape)*1
event_set_Type.append(event_Type)
event_Loc  = np.ones(event.shape)*0
event_Loc[[0]] = 1
event_set_Loc.append(event_Loc)

event = np.array([
64329,64368,67726,68553,79693,82461,89519,100169,101903,118211,125764
])
event_set.append(event) #D2
event_Type = np.ones(event.shape)*1
event_Type[[7]] = 0
event_set_Type.append(event_Type)
event_Loc  = np.ones(event.shape)*0
event_set_Loc.append(event_Loc)

exp_id = 8
experiment_mscore[fileloc_set[exp_id]] = event_set
experiment_info['events'] = event_set
experiment_info['events type'] = event_set_Type
experiment_info['events loc']  = event_set_Loc
experiment_info['sex'] = [0,1,1,1,0,0,0,0]
experiment_info['genotype'] = [0,2,2,2,2,2,2,2]
experiment_info['starting_time'] = [1595,1675,2198,3059,3983,4192,4369,4677]
experiments_mscore[fileloc_set[exp_id]] = experiment_info

# 07-23-2024-exp1

experiment_info = {}
event_set = []
event_set_Type = []
event_set_Loc = []
event = np.array([
    23731,25332,26301,27106,47859,51563,65008,65178,68180,90159,92473,123198
])
event_set.append(event) #A1

event_Type = np.ones(event.shape)*2
event_set_Type.append(event_Type)
event_Loc = np.ones(event.shape)*1
event_Loc[0] = 0
event_Loc[[3,10]] = 2
event_set_Loc.append(event_Loc)




event = np.array([
    10814,11268,11674,12205,13074,13453,14272,15682,16168,17253,17598,17654,18496,18871,19139,20514,22354
])
event_set.append(event) #B1

event_Type = np.ones(event.shape)*2
event_set_Type.append(event_Type)
event_Loc = np.ones(event.shape)*1
event_set_Loc.append(event_Loc)



event = np.array([
    11809,12556,12555,13009,20389,33962,37778,39947,65547,94877,95586,103342,104305,119424
])
event_set.append(event) #C1

event_Type = np.ones(event.shape)*2
event_set_Type.append(event_Type)
event_Loc = np.ones(event.shape)*1
event_Loc[[7]]     = 0
event_Loc[[5,9,13]] = 2
event_set_Loc.append(event_Loc)


event = np.array([
    29472,29500,63510,63674,71790,77436,89204,93734,96581,96612,99942,99983,101247,116232,117422,117426
])
event_set.append(event) #D1

event_Type = np.ones(event.shape)*1
event_set_Type.append(event_Type)
event_Loc = np.ones(event.shape)*0
event_set_Loc.append(event_Loc)



event = np.array([
    17463,27169,32879,33213,37089,39403,39468,42958,43537,47983,48027,79701,79724,79822,79951
])
event_set.append(event) #A2

event_Type = np.ones(event.shape)*1
event_Type[[0,7,9,14]] = 0 
event_set_Type.append(event_Type)
event_Loc = np.ones(event.shape)*0
event_set_Loc.append(event_Loc)

event = np.array([])
event_set.append(event) #B2
event_Type = np.array([])
event_set_Type.append(event_Type)
event_Loc = np.array([])
event_set_Loc.append(event_Loc)

event = np.array([])
event_set.append(event) #C2
event_Type = np.array([])
event_set_Type.append(event_Type)
event_Loc = np.array([])
event_set_Loc.append(event_Loc)

event = np.array([])
event_set.append(event) #D2
event_Type = np.array([])
event_set_Type.append(event_Type)
event_Loc = np.array([])
event_set_Loc.append(event_Loc)

exp_id = 9

experiment_info['events'] = event_set
experiment_info['events type'] = event_set_Type
experiment_info['events loc']  = event_set_Loc
experiment_info['sex']      = [1,-1,1,0,0,-1,-1,0]
experiment_info['genotype'] = [0,-1,0,2,0,-1,-1,1]
experiment_info['starting_time'] = [1813,2064,2332,2537,3088,3258,3878,4098]
experiment_mscore[fileloc_set[exp_id]] = event_set
experiments_mscore[fileloc_set[exp_id]] = experiment_info

# 07-23-2024-exp2

experiment_info = {}
event_set = []
event_set_Type = []
event_set_Loc = []

event = np.array([
    12500,19321,44635,121423,121979,
])
event_set.append(event) #A1

event_Type = np.ones(event.shape)*2
event_Type[[0,1,2]] = 0
event_set_Type.append(event_Type)
event_Loc = np.ones(event.shape)*0
event_Loc[4] = 1
event_set_Loc.append(event_Loc)



event = np.array([
    
])
event_set.append(event) #B1

event_Type = np.array([])
event_set_Type.append(event_Type)
event_Loc = np.array([])
event_set_Loc.append(event_Loc)


event = np.array([
    28267,33420,33877,44910,65962,70440,72417,99302,107165,
])
event_set.append(event) #C1

event_Type = np.ones(event.shape)*2
event_Type[[4,5,6,7,8]] = 3
event_set_Type.append(event_Type)
event_Loc = np.ones(event.shape)*0
event_set_Loc.append(event_Loc)


event = np.array([
    53534,53747,58106,58162,58175,58828,77479,77639,78159,79418,79518,80104,80292,92447,92512,92891,
101500,101513,101531,107497,107545,107580,107696,116404,116607,119885,122334,122404,122435,125528
])
event_set.append(event) #D1
event_Type = np.ones(event.shape)*3
event_set_Type.append(event_Type)
event_Loc = np.ones(event.shape)*0
event_set_Loc.append(event_Loc)


event = np.array([
    47761,48563,49241,57025,71833,72193,72269,72634,72783,82708,83357,90848,90901,91319,94931,95150,
    97527,114523,114527,114530,114534,114554,114796,115825,116366,116436,116736,122222,122284,124650
])
event_set.append(event) #A2

event_Type = np.ones(event.shape)*1
event_Type[[6,15,17,18,27]] = 0
event_set_Type.append(event_Type)
event_Loc = np.ones(event.shape)*0
event_set_Loc.append(event_Loc)


event = np.array([
    45308,45323,45388,55464,96173,109433,116453,116484,116714,116749
])
event_set.append(event) #B2

event_Type = np.ones(event.shape)*1
event_set_Type.append(event_Type)
event_Loc = np.ones(event.shape)*0
event_Loc[[3,4]] = 1
event_set_Loc.append(event_Loc)


event = np.array([
    23575,23602,23606,23612,23651,25447,25562,25565,25569,25582,28212,30995,34249,34940,39338,39373,39338,39373,
40542,40792,40937,43109,43984,45407,47446,47455,47863,55539,55685,57927,62023,66481,68065,71426,71472,76266,76279,
76290,76348,76931,77580,80573,90820,93495,94675,95802,95806,98907,99291,100399,100591,101751,101921,102448,106334,
106367,108435,108576,108641,109780,111245,112013,120926,120937,123404,123731,123792,124883,124884,125636
])
event_set.append(event) #C2

event_Type = np.ones(event.shape)*1
event_Type[[22,60]] = 0
event_set_Type.append(event_Type)
event_Loc = np.ones(event.shape)*0
event_set_Loc.append(event_Loc)


event = np.array([
    
])
event_set.append(event) #D2
event_Type = np.array([])
event_set_Type.append(event_Type)
event_Loc = np.array([])
event_set_Loc.append(event_Loc)

exp_id = 10

experiment_info['events'] = event_set
experiment_info['events type'] = event_set_Type
experiment_info['events loc']  = event_set_Loc
experiment_info['sex']      = [1,-1,1,1,0,0,0,-1]
experiment_info['genotype'] = [0,-1,0,2,0,0,0,-1]
experiment_info['starting_time'] = [1756, 2319, 2544, 2859, 3450, 3980, 4282, 0]
experiment_mscore[fileloc_set[exp_id]] = event_set
experiments_mscore[fileloc_set[exp_id]] = experiment_info

# 07-24-2024-exp1

experiment_info = {}
event_set = []
event_set_Type = []
event_set_Loc = []

event = np.array([
    9590,30242,33279,34428,34511,34800,43814,45106,46010,49533,52542,53040,56306,59125,62561,69123,72878,73245,
    73409,73636,74330,75974,80273,84199,85970,86340,91397,96344,97924,98514,101613,102293,104411,104646,108318,
    112451,116236,116450,116859,120045,123759,
])
event_set.append(event) #A1
event_Type = np.ones(event.shape)*2
event_Type[[36,37]] = 3
event_set_Type.append(event_Type)
event_Loc = np.ones(event.shape)*0
event_Loc[[0,2,3,4,5,7,8,10,11,13]] = 1
event_Loc[[1,9]] = 2
event_set_Loc.append(event_Loc)


event = np.array([
    9922,59707,105911,118276,118313,118329
])
event_set.append(event) #B1

event_Type = np.ones(event.shape)*1
event_Type[[3,4]] = 0
event_set_Type.append(event_Type)
event_Loc = np.ones(event.shape)*0
event_Loc[0] = 1
event_set_Loc.append(event_Loc)


event = np.array([
    104388,120219
])
event_set.append(event) #C1
event_Type = np.ones(event.shape)*1
event_set_Type.append(event_Type)
event_Loc = np.ones(event.shape)*0
event_set_Loc.append(event_Loc)



event = np.array([
    17018,18509,19101,22530
])
event_set.append(event) #D1
event_Type = np.ones(event.shape)*2
event_Type[3] = 0
event_set_Type.append(event_Type)
event_Loc = np.ones(event.shape)*0
event_set_Loc.append(event_Loc)


event = np.array([
    9223,10552,15324,15977,20969,21691,25804,26551,27432,34934,39167,40553,41106,41695,42029,42434,43773,44362,
45312,47382,47913,48873,50432,63293,67139,69743,73786,75000,81044,86284,86893,90498,93261,99874,104099,111537,
111801
])
event_set.append(event) #A2
event_Type = np.ones(event.shape)*2
event_Type[[10,11,19]] = 0
event_Type[[24,31]] = 3
event_set_Type.append(event_Type)
event_Loc = np.ones(event.shape)*1
event_Loc[[12,13,14,15,16,19,26,28]] = 2
event_Loc[[10,11,29,34,35,36]] = 0
event_set_Loc.append(event_Loc)

event = np.array([
    49569,49833,58126,58155,60928,64547,64579,74042,80667,80723,81362,81447,83311,83317,83371,98788,
98827,99605,99624,100732,100767,106872,110730,115876,115905,115982,119333,122664
])
event_set.append(event) #B2
event_Type = np.ones(event.shape)*1
event_Type[15] = 0
event_set_Type.append(event_Type)
event_Loc = np.ones(event.shape)*0
event_set_Loc.append(event_Loc)

event = np.array([
    57715,57725,63072,93327,93340,94515,98256,98850,103063,103239,104266,115339,115349,118505,118781,119057,
119738,122295,122374,123803,123825
])
event_set.append(event) #C2
event_Type = np.ones(event.shape)*1
event_Type[2] = 0
event_set_Type.append(event_Type)
event_Loc = np.ones(event.shape)*0
event_set_Loc.append(event_Loc)

event = np.array([
    13833
])
event_set.append(event) #D2
event_Type = np.ones(event.shape)*2
event_set_Type.append(event_Type)
event_Loc = np.ones(event.shape)*0
event_set_Loc.append(event_Loc)

exp_id = 11

experiment_info['events'] = event_set
experiment_info['events type'] = event_set_Type
experiment_info['events loc']  = event_set_Loc
experiment_info['sex']      = [1,0,0,1,1,0,0,1]
experiment_info['genotype'] = [0,0,0,1,0,2,2,1]
experiment_info['starting_time'] = [2189,2552,2838,3030,5189,4637,4285,3775]
experiment_mscore[fileloc_set[exp_id]] = event_set
experiments_mscore[fileloc_set[exp_id]] = experiment_info

# 07-24-2024-exp2

experiment_info = {}
event_set = []
event_set_Type = []
event_set_Loc = []

event = np.array([
    80226,86278,86973,89502,89535,114681,118879
])
event_set.append(event) #A1
event_Type = np.ones(event.shape)*3
event_set_Type.append(event_Type)
event_Loc = np.ones(event.shape)*0
event_set_Loc.append(event_Loc)


event = np.array([
    
])
event_set.append(event) #B1
event_Type = np.array([])
event_set_Type.append(event_Type)
event_Loc = np.array([])
event_set_Loc.append(event_Loc)

event = np.array([
    
])
event_set.append(event) #C1
event_Type = np.array([])
event_set_Type.append(event_Type)
event_Loc = np.array([])
event_set_Loc.append(event_Loc)


event = np.array([
    81626,86499,86549,101873,101981,102052,102890,103078,103217,104773,104820,104968,108398,108431,108571,108662,
108815,108855,114424,118355
])
event_set.append(event) #D1
event_Type = np.ones(event.shape)*1
event_set_Type.append(event_Type)
event_Loc = np.ones(event.shape)*0
event_set_Loc.append(event_Loc)


event = np.array([
    25022,41099,41136,42360,49748,49774,58247,58261,58477,58980,62658,62710,66600,66693,67066,69724,72972,77818,
82595,97801,97808,98401,102554,103196,105503,105524,110839,111535,111996,112291,122420
])
event_set.append(event) #A2
event_Type = np.ones(event.shape)*1
event_set_Type.append(event_Type)
event_Loc = np.ones(event.shape)*0
event_set_Loc.append(event_Loc)

event = np.array([
    14262,14285,14370,36403,36469,36828,38305,43700,44067,44278,44357,46446,46470,49392,50550,50557,
55604,55772,59381,60717,61299,65226,65228,66312,68828,68833,70250,74746,74749,77754,79210,79621,
79704,80860,82060,85306,85475,87204,88241,90624,92720,98030,105014,105676,107684,107716,107800,109561,
110375,110927,110976,112492,114541,114958,116401,119695,120393,121764,123641
])
event_set.append(event) #B2
event_Type = np.ones(event.shape)*1
event_set_Type.append(event_Type)
event_Loc = np.ones(event.shape)*0
event_set_Loc.append(event_Loc)

event = np.array([
    27678,41311,45136,55596,65045,65306,97337
])
event_set.append(event) #C2
event_Type = np.ones(event.shape)*1
event_set_Type.append(event_Type)
event_Loc = np.ones(event.shape)*0
event_set_Loc.append(event_Loc)

event = np.array([
    25143,25686,27260,27298,27325,39630,39650,46943,47715,62139,71416,71751,89484,97029,98423,102431,
102454,102471,102942,106133,109034
])
event_set.append(event) #D2
event_Type = np.ones(event.shape)*1
event_set_Type.append(event_Type)
event_Loc = np.ones(event.shape)*0
event_set_Loc.append(event_Loc)


exp_id = 12

experiment_info['events'] = event_set
experiment_info['events type'] = event_set_Type
experiment_info['events loc']  = event_set_Loc
experiment_info['sex']      = [1,-1,-1,0,0,0,0,0]
experiment_info['genotype'] = [2,-1,-1,1,2,0,1,1]
experiment_info['starting_time'] = [1662,1885,2048,2351,3514,3169,2958,2701]
experiment_mscore[fileloc_set[exp_id]] = event_set
experiments_mscore[fileloc_set[exp_id]] = experiment_info

# 07-24-2024-exp3

experiment_info = {}
event_set = []
event_set_Type = []
event_set_Loc = []

event = np.array([
    22679,22710,27059,27083,34681,34740,40625,56453,66056,71078,71317,77863,85414,86266,86360,
89131,89475,104044,106990,113251,114233,117552,120806,120858,120892,121570,123159
])
event_set.append(event) #A1
event_Type = np.ones(event.shape)*1
event_set_Type.append(event_Type)
event_Loc = np.ones(event.shape)*0
event_set_Loc.append(event_Loc)

event = np.array([
    124106
])
event_set.append(event) #B1
event_Type = np.ones(event.shape)*1
event_set_Type.append(event_Type)
event_Loc = np.ones(event.shape)*0
event_set_Loc.append(event_Loc)

event = np.array([
    28267,43310,62453,74009,74043,74098,74707,76357,76798,77260,77353,93611,98852,99006,99070,
99375,99465,100382,104458,116958,117041,122156
])
event_set.append(event) #C1
event_Type = np.ones(event.shape)*1
event_set_Type.append(event_Type)
event_Loc = np.ones(event.shape)*0
event_set_Loc.append(event_Loc)

event = np.array([
    
])
event_set.append(event) #D1
event_Type = np.array([])
event_set_Type.append(event_Type)
event_Loc = np.array([])
event_set_Loc.append(event_Loc)

event = np.array([
    50696,50719,50731,50743,50754,50915,50926,50938,50955,50976,50993,51009,51064,51650,51659,51738,
51798,52036,52060,52218,52266,65528,65563,66361,68470,68539,68599,68616,68631,68702,68845,68989,69362,72060,
72764,80390, 84351,88261,88299,88557,90581,90671,92849,92857,101591,103493,109857,110595
])
event_set.append(event) #A2
event_Type = np.ones(event.shape)*3
event_Type[31] = 2
event_set_Type.append(event_Type)
event_Loc = np.ones(event.shape)*0
event_set_Loc.append(event_Loc)

event = np.array([
    21812,22923,24905,39401,44515,44742,51687,51696,51716,51751,51803,51816,51907,51966,52021,52048,52057,
52121,58776,60310,67946,71878,73134,78066,78787,85186,94518
])
event_set.append(event) #B2
event_Type = np.ones(event.shape)*3
event_Type[0] = 0
event_Type[19] = 2
event_set_Type.append(event_Type)
event_Loc = np.ones(event.shape)*0
event_Loc[[0,1,2,4,5,19,20,21,23,24,25,26]] = 1
event_set_Loc.append(event_Loc)

event = np.array([
    
])
event_set.append(event) #C2
event_Type = np.array([])
event_set_Type.append(event_Type)
event_Loc = np.array([])
event_set_Loc.append(event_Loc)
event = np.array([
    
])
event_set.append(event) #D2
event_Type = np.array([])
event_set_Type.append(event_Type)
event_Loc = np.array([])
event_set_Loc.append(event_Loc)


exp_id = 13

experiment_info['events'] = event_set
experiment_info['events type'] = event_set_Type
experiment_info['events loc']  = event_set_Loc
experiment_info['sex']      = [0,0,0,0,1,1,1,-1]
experiment_info['genotype'] = [2,2,1,1,2,2,1,-1]
experiment_info['starting_time'] = [1148,1444,2092,2473,4230,3919,3296,3305]
experiment_mscore[fileloc_set[exp_id]] = event_set
experiments_mscore[fileloc_set[exp_id]] = experiment_info

# 07-24-2024-exp4

experiment_info = {}
event_set = []
event_set_Type = []
event_set_Loc = []

event = np.array([
    
])
event_set.append(event) #A1
event_Type = np.array([])
event_set_Type.append(event_Type)
event_Loc = np.array([])
event_set_Loc.append(event_Loc)
event = np.array([
    
])
event_set.append(event) #B1
event_Type = np.array([])
event_set_Type.append(event_Type)
event_Loc = np.array([])
event_set_Loc.append(event_Loc)
event = np.array([
    
])
event_set.append(event) #C1
event_Type = np.array([])
event_set_Type.append(event_Type)
event_Loc = np.array([])
event_set_Loc.append(event_Loc)
event = np.array([
    
])
event_set.append(event) #D1
event_Type = np.array([])
event_set_Type.append(event_Type)
event_Loc = np.array([])
event_set_Loc.append(event_Loc)
event = np.array([
    120123,120649,121946,121986,122030
])
event_set.append(event) #A2
event_Type = np.ones(event.shape)*3
event_Type[0] = 2
event_set_Type.append(event_Type)
event_Loc = np.ones(event.shape)*0
event_set_Loc.append(event_Loc)


event = np.array([
    
])
event_set.append(event) #B2
event_Type = np.array([])
event_set_Type.append(event_Type)
event_Loc = np.array([])
event_set_Loc.append(event_Loc)
event = np.array([
    7469,7534,7597,10082,20962,21024,21038,21101,36353,36392,55862,56671,56856,59953,60144,60151,61202,61373,
61670,61957,62120,62165,62367,62425,65399,65452,65570,65680,68318,68403,68419,68907,71887,73124,73288,77239,
77820,78453,78610,79478,85447,85739,87224,87314,87505,97194,97470,101063,101299,102400,102466,116043,118479,
118780,121317
])
event_set.append(event) #C2
event_Type = np.ones(event.shape)*1
event_set_Type.append(event_Type)
event_Loc = np.ones(event.shape)*0
event_Loc[[0,1,2,3]] = 1
event_set_Loc.append(event_Loc)

event = np.array([
    
])
event_set.append(event) #D2
event_Type = np.array([])
event_set_Type.append(event_Type)
event_Loc = np.array([])
event_set_Loc.append(event_Loc)

exp_id = 14

experiment_info['events'] = event_set
experiment_info['events type'] = event_set_Type
experiment_info['events loc']  = event_set_Loc
experiment_info['sex']      = [0,-1,-1,-1,1,1,0,-1]
experiment_info['genotype'] = [1,-1,-1,-1,1,1,1,-1]
experiment_info['starting_time'] = [2244,0,0,0,3136,3134,2839,0]
experiment_mscore[fileloc_set[exp_id]] = event_set
experiments_mscore[fileloc_set[exp_id]] = experiment_info

np.save('experiment_mscore.npy', experiment_mscore)
np.save('experiments_mscore.npy', experiments_mscore)

for i in range(15):
    print(fileloc_set[i])
    experiment = experiments_mscore[fileloc_set[i]]
    print(experiment['sex'])
    print(experiment['genotype'])
