==> DataStreamLine/03.copy_selected_original_video.ipynb <==
import numpy as np
import pandas as pd
from pathlib import Path
import os
import matplotlib.pyplot as plt

import scipy.signal as signal
import imageio as io
from tqdm import tqdm
import cv2
import shutil

import sys
sys.path.insert(1, '../numphly')
import numphly as nly

dirloc = '../../../../../../Documents/'

fileloc_set   = np.load('experiments_directory.npy')
num_file = len(fileloc_set)
videofile_set = np.load('experiments_well_setup.npy')
num_well = len(videofile_set)
experiments_index = np.load('experiments_index.npy')

experiments_mscore = np.load('experiments_mscore.npy',allow_pickle='TRUE').item()
sex_name      = ['F','M']
genotype_name = ['RxT', 'RxW', 'TxW']

frame_rate = 60
frame_n = 126000
height = 576 
width = 576
w_o = 576
h_o = w_o

output_fileloc = '../../FlyBehaviorSet/'
OG_recording_folder= 'original_recording_for_tracking/R72'
converted_folder = 'Converted/R72_videos/'
for i in range(num_file):
    experiment = experiments_mscore[fileloc_set[i]]
    sexes = experiment['sex']
    gtypes = experiment['genotype']
    experiment_index = experiments_index[i]
    for j in range(8):
        well = videofile_set[j]
        output_file_prefix = 'v_'+experiment_index+'_'+well
        if sexes[j] != -1:
#             # move usable OG recordings to a single folder
            video   = dirloc+fileloc_set[i]+'well_'+videofile_set[j]+'.mp4'
            PATH_FOI = output_fileloc + OG_recording_folder
#             shutil.copyfile(video, PATH_FOI+'/'+output_file_prefix+'.mp4')
#             print(output_file_prefix+'.mp4')
            #R72A10xTRPA1 female copy
            if sexes[j] == 0 and gtypes[j] == 0:
                video   = output_fileloc+converted_folder+fileloc_set[i]+'well_'+videofile_set[j]+'_ts.mp4'
                PATH_FOI = output_fileloc + converted_folder
                shutil.copyfile(video, PATH_FOI+output_file_prefix+'_ts.mp4')
                print(video)


output_fileloc = '../../FlyBehaviorSet/'
OG_recording_folder= 'original_recording_for_tracking/R72'
converted_folder = 'Converted/R72_videos/'
for i in range(num_file):
    experiment = experiments_mscore[fileloc_set[i]]
    sexes = experiment['sex']
    gtypes = experiment['genotype']
    experiment_index = experiments_index[i]
    for j in range(8):
        well = videofile_set[j]
        output_file_prefix = 'v_'+experiment_index+'_'+well
        if sexes[j] != -1:
#             # move usable OG recordings to a single folder

            #R72A10xTRPA1 female copy
#             if sexes[j] == 0 and gtypes[j] == 0:
#                 video   = output_fileloc+converted_folder+fileloc_set[i]+'well_'+videofile_set[j]+'_ts.mp4'
#                 PATH_FOI = output_fileloc + converted_folder
#                 shutil.copyfile(video, PATH_FOI+output_file_prefix+'_ts.mp4')
#                 print(video)
            #R72A10xW1118 female copy
            if sexes[j] == 0 and gtypes[j] == 1:
                video   = output_fileloc+converted_folder+fileloc_set[i]+'well_'+videofile_set[j]+'_ts.mp4'
                PATH_FOI = output_fileloc + converted_folder
#                 shutil.copyfile(video, PATH_FOI+output_file_prefix+'_ts.mp4')
                print(output_file_prefix)

# R72 clip a section of orginal video to test programs. Original videos size is large and unwieldy

OG_recording_folder_part = 'original_recording_for_tracking/R72_part'
# fourcc = cv2.VideoWriter_fourcc(*'FFV1')
fourcc = cv2.VideoWriter_fourcc('m', 'p', '4', 'v')
frame_n_part = frame_rate*60*5

# for i in range(1):
# for i in range(7):
#     experiment = experiments_mscore[fileloc_set[i]]
#     sex = experiment['sex']
#     experiment_index = experiments_index[i]
#     for j in range(8):
#         well = videofile_set[j]
#         output_file_prefix = 'v_'+experiment_index+'_'+well
#         if sex[j] != -1:
# #             # move usable OG recordings to a single folder
#             video    = output_fileloc + OG_recording_folder +'/'+output_file_prefix+'.mp4'
#             video_out= output_fileloc + OG_recording_folder_part +'/'+output_file_prefix+'.mp4'
#             cap = cv2.VideoCapture(video)
#             out = cv2.VideoWriter(video_out, fourcc, frame_rate, (height,width))
#             for f in range(frame_n-frame_n_part, frame_n):
#                 cap.set(cv2.CAP_PROP_POS_FRAMES, f)
#                 res, frame = cap.read()
#                 out.write(frame)
                
#             cap.release()
#             out.release()
#             cv2.destroyAllWindows()
#             print(video_out)

print(frame_n-frame_n_part)
