==> DataStreamLine/04.crop_parameters.ipynb <==
import numpy as np
import pandas as pd
from pathlib import Path
import os
import matplotlib.pyplot as plt

import scipy.signal as signal
import imageio as io
from tqdm import tqdm
import cv2
import shutil

import sys
sys.path.insert(1, '../numphly')
import numphly as nly

dirloc = '../../../../../../Documents/'
fileloc_set   = np.load('experiments_directory.npy')
num_file = len(fileloc_set);
videofile_set = np.load('experiments_well_setup.npy')
experiments_index = np.load('experiments_index.npy')
print(fileloc_set, len(fileloc_set))
print("")
print(videofile_set)
print("")
print(experiments_index)
sex_name      = ['F','M']
genotype_name = ['RxT', 'RxW', 'TxW']
frame_rate = 60

experiments_mscore = np.load('experiments_mscore.npy',allow_pickle='TRUE').item()
experiments_mscore_1 = np.load('experiments_mscore.npy',allow_pickle='TRUE').item()
cropped_recording_folder= 'cropped_recording_for_tracking'
output_fileloc = '../../FlyBehaviorSet/'
OG_recording_folder= 'original_recording_for_tracking'
converted_folder = 'Converted/R72_videos/'


first_frame = 20000
last_frame = 125000
common_difference = frame_rate*30
event = np.arange(first_frame,last_frame, common_difference)
np.set_printoptions(precision=1)
{}
ii = 0
for i in range(num_file):
# for i in range(len(experiments_mscore)):
# for i in range(1,num_file):
    experiment = experiments_mscore[fileloc_set[i]]
    sexes = experiment['sex']
    gtypes = experiment['genotype']
    event_set_circle = []
    if "events" in experiment:
        events = experiment['events']
        experiment_index = experiments_index[i]
        print(experiment_index)
        for j in range(8):
#         for j in range(1):
            well_arena = np.empty((0,3))
            well_food  = np.empty((0,3))
            well = videofile_set[j]
            sex = sexes[j]
            gtype = gtypes[j]
            output_file_prefix = 'v_'+experiment_index+'_'+well
            if sex != -1:
                # clipping videos for labeling and move them into a single directory
                video   = output_fileloc + OG_recording_folder +'/R72/'+output_file_prefix+'.mp4'
                
#                 video   = output_fileloc+converted_folder+fileloc_set[i]+'well_'+videofile_set[j]+'_ts.mp4'

                cap = cv2.VideoCapture(video)
                length = int(cap.get(cv2.CAP_PROP_FRAME_COUNT))
                n_circles = 0
                for region in event:
                    cap.set(cv2.CAP_PROP_POS_FRAMES, region)
                    res, frame = cap.read()
                    gray = cv2.cvtColor(frame, cv2.COLOR_BGR2GRAY)
                    gray_blurred = cv2.medianBlur(gray, 5)
                    arena = cv2.HoughCircles(gray_blurred, cv2.HOUGH_GRADIENT, 1, 20, 
                                               param1=50, param2=50, minRadius=225, maxRadius=245)
#                         food = cv2.HoughCircles(gray_blurred, cv2.HOUGH_GRADIENT, 1, 20, 
#                                                    param1=55, param2=55, minRadius=50, maxRadius=100)
                    if arena is not None:
                        well_arena = np.vstack([well_arena, (arena[0,:])[0,:]])
                        n_circles = n_circles+1
                    if n_circles >= 50:
                        break
                cap.release()
                
                cv2.destroyAllWindows()
                circle = np.mean(well_arena, axis = 0)
                circle = np.round(circle).astype("int")
                print(output_file_prefix, circle, well_arena.shape)
                cv2.circle(frame, (circle[0], circle[1]), 240, (0, 255, 0), 2)
                # Draw the center of the circle
                cv2.circle(frame, (circle[0], circle[1]), 2, (0, 0, 255), 3)
                image   = output_fileloc + cropped_recording_folder +'/'+output_file_prefix+'_radius.png'
                cv2.imwrite(image, frame, [cv2.IMWRITE_PNG_COMPRESSION, 0])
                ii += 1
            event_set_circle.append(well_arena)
    else:
        print('no events in experiments_mscore')
    
    experiment['arena circle']= event_set_circle
    experiments_mscore[fileloc_set[i]] = experiment
print('done',ii)




# imgplot = plt.imshow(frame)
# plt.show()
# print(circles.shape, circles)

print(circle)

np.set_printoptions(precision=1)
crop_h = []
# for i in range(num_file):
# for i in range(2):
for i in range(len(experiments_mscore)):
    experiment   = experiments_mscore[fileloc_set[i]]
    arena_circle = experiment['arena circle']
    sexes = experiment['sex']
    if "events" in experiment:
        for j in range(8):
            sex = sexes[j]
            well = videofile_set[j]
            well_arena   = arena_circle[j]

            if sex != -1:
                d_std= np.sum(np.std(well_arena, axis = 0))
                circle = np.mean(well_arena, axis = 0)
                radius = circle[2]#+d_std
                center = circle[0:2]
                corner1 = circle[0:2]-radius
                corner2 = 576-(circle[0:2]+radius)


                
                

                
                
                
                if corner1[0]>0 and corner1[1]>0 and corner2[0]>0 and corner2[1]>0:
                    crop_h.append(circle)
                    print(center, radius, circle.shape)
                else:
                    print(fileloc_set[i],well,center,'wait a minute', radius)# print(crop_h.shape)
                    
# np.save('crop_h.npy', crop_h)
# np.save('experiments_mscore.npy',experiments_mscore)

def detect_circles(image_path):
    """
    Detects circles in an image using the Hough Circle Transform.

    Args:
        image_path: Path to the image file.

    Returns:
        A list of detected circles, where each circle is represented by 
        a tuple of (x, y, radius).
    """

    # Load the image
    img = cv2.imread(image_path)

    # Convert to grayscale
    gray = cv2.cvtColor(img, cv2.COLOR_BGR2GRAY)

    # Blur the image to reduce noise
    gray_blurred = cv2.medianBlur(gray, 5)

    # Detect circles using Hough Circle Transform
    circles = cv2.HoughCircles(gray_blurred, cv2.HOUGH_GRADIENT, 1, 20, 
                              param1=50, param2=30, minRadius=0, maxRadius=0)

    # Ensure at least some circles were found
    if circles is not None:
        # Convert the (x, y) coordinates and radius of the circles to integers
        circles = np.round(circles[0, :]).astype("int")

        # Draw detected circles on the original image
        for (x, y, r) in circles:
            # Draw the outer circle
            cv2.circle(img, (x, y), r, (0, 255, 0), 2)
            # Draw the center of the circle
            cv2.circle(img, (x, y), 2, (0, 0, 255), 3)

    # Display the resulting frame
    cv2.imshow('Detected Circles', img)
    cv2.waitKey(0)
    cv2.destroyAllWindows()

    return circles

# # Example usage:
# image_path = "path/to/your/image.jpg"  # Replace with the actual image path
# detected_circles = detect_circles(image_path)

# if detected_circles is not None:
#     print("Detected circles:", detected_circles)
# else:
#     print("No circles detected.")


first_frame = 1000
last_frame = 125000
common_difference = frame_rate*30
event = np.arange(first_frame,last_frame, common_difference)
np.set_printoptions(precision=1)


for i in range(num_file):
# for i in range(len(experiments_mscore)):
# for i in range(1,num_file):
    experiment = experiments_mscore[fileloc_set[i]]
    sexes = experiment['sex']
    gtypes = experiment['genotype']
    event_set_circle = []
    if "events" in experiment:
        events = experiment['events']
        experiment_index = experiments_index[i]
        print(experiment_index)
        for j in range(8):
#         for j in range(1):
            well_arena = np.empty((0,3))
            well_food  = np.empty((0,3))
            well = videofile_set[j]
            sex = sexes[j]
            gtype = gtypes[j]
            output_file_prefix = 'v_'+experiment_index+'_'+well
            if sex != -1:
                # clipping videos for labeling and move them into a single directory
                video   = output_fileloc + OG_recording_folder +'/R72/'+output_file_prefix+'.mp4'
                print(video)
                cap = cv2.VideoCapture(video)
                length = int(cap.get(cv2.CAP_PROP_FRAME_COUNT))
                n_circles = 0
                for region in event:
                    cap.set(cv2.CAP_PROP_POS_FRAMES, region)
                    res, frame = cap.read()
                    gray = cv2.cvtColor(frame, cv2.COLOR_BGR2GRAY)
                    gray_blurred = cv2.medianBlur(gray, 5)
                    arena = cv2.HoughCircles(gray_blurred, cv2.HOUGH_GRADIENT, 1, 20, 
                                               param1=50, param2=50, minRadius=220, maxRadius=250)
#                         food = cv2.HoughCircles(gray_blurred, cv2.HOUGH_GRADIENT, 1, 20, 
#                                                    param1=55, param2=55, minRadius=50, maxRadius=100)
                    if arena is not None:
                        well_arena = np.vstack([well_arena, (arena[0,:])[0,:]])
                        n_circles = n_circles+1
                    if n_circles >= 20:
                        break
                cap.release()
                
                cv2.destroyAllWindows()
                circle = np.mean(well_arena, axis = 0)
                circle = np.round(circle).astype("int")

                cv2.circle(frame, (circle[0], circle[1]), 240, (0, 255, 0), 2)
                # Draw the center of the circle
                cv2.circle(frame, (circle[0], circle[1]), 2, (0, 0, 255), 3)
                image   = output_fileloc + cropped_recording_folder +'/'+output_file_prefix+'.png'
                cv2.imwrite(image, frame, [cv2.IMWRITE_PNG_COMPRESSION, 0])
            event_set_circle.append(well_arena)
    else:
        print('no events in experiments_mscore')
    
    experiment['arena circle']= event_set_circle
    experiments_mscore[fileloc_set[i]] = experiment
print('done')




# imgplot = plt.imshow(frame)
# plt.show()
# print(circles.shape, circles)




for i in range(num_file):
    experiment = experiments_mscore[fileloc_set[i]]
    sexes = experiment['sex']
    gtypes = experiment['genotype']
    event_set_circle = []
    if "events" in experiment:
        events = experiment['events']
        events_type  = experiment['events type']
        experiment_index = experiments_index[i]
        for j in range(8):
#         for j in range(1):
            well_arena = np.empty((0,3))
            well_food  = np.empty((0,3))
            well = videofile_set[j]
            sex = sexes[j]
            gtype = gtypes[j]
            events_type_j = events_type[j]
            output_file_prefix = 'v_'+experiment_index+'_'+well
            if sex != -1:
                if sex == 0:
                    if gtype == 2:
#                         print(output_file_prefix)
                        n_event = len(events_type_j)
                        if n_event == 0:
                            print(0)
                        else:
                            print(len(events_type_j[events_type_j==1]))
                        

    else:
        print('no events in experiments_mscore')

print('done')
