==> DataStreamLine/11.downsampling.ipynb <==
import pandas as pd
from pathlib import Path
import numpy as np
import os
import matplotlib.pyplot as plt
from matplotlib.lines import Line2D
from matplotlib.patches import Patch
import matplotlib.ticker as mtick
import scipy.signal as signal
import imageio as io
from tqdm import tqdm
import pickle, re
import cv2
font = cv2.FONT_HERSHEY_SIMPLEX
rgb = (255, 255, 255)

import sys
sys.path.insert(1, '../numphly')
import numphly as nly
import mcam_tools
import tracker_decoder as tkr_dec
import warnings
warnings.filterwarnings("ignore", category=RuntimeWarning)


def check_characters(input_string, target_characters):
    return set(target_characters).issubset(set(input_string))


exp_target = 'R72'
gtype_name = ['R72A10xTRPA1','R72A10xW1118','W1118xTRPA1']
sex_name = ['female','male']
h5_dir = '../../FlyBehaviorSet/DLC_tracking/'+exp_target
output_fileloc = '../../FlyBehaviorSet/'
OG_recording_folder      = 'original_recording_for_tracking/'+exp_target
converted_folder         = 'Converted/R72_videos/'
cropped_recording_folder = 'cropped_recording_for_tracking'  
dir_MCAM = '../DatastreamLine_MCAM_Tracking/'
shared_data_folder = '../Shared_data/'

scorer_floor  = 'DLC_resnet152_floor_topview_12label_1uniqueJan4shuffle1_100000'
# scorer_wall   = 'DLC_resnet152_wall_sideview_7labelsJan1shuffle1_100000'
scorer_wall   = 'DLC_resnet152_wall_sideview_7labelsJan1shuffle1_90000'
# scorer_cross  = 'DLC_resnet152_wall_cross_section_view_8labelFeb19shuffle1_100000'
scorer_cross  = 'DLC_resnet152_wall_cross_section_view_8labelFeb19shuffle1_30000'


pcutoff = 0
frame_n = 126000
frame_rate = 60
frame_n_e = frame_rate*60*30

color_BP = [(0,0,255),(0,255,0),(255,0,0)]
arena_parameters = {
    'well_radius': 8.4,  # mm
    'food_cup_radius': 3.5,  # mm
}

experiments_index  = np.load('experiments_index.npy')
fileloc_set        = np.load('experiments_directory.npy')
num_file = len(fileloc_set)
videofile_set      = np.load('experiments_well_setup.npy')
num_well = len(videofile_set)
experiments_mscore = np.load('experiments_mscore.npy',allow_pickle='TRUE').item()
crop_h = np.load('crop_h.npy')
print(len(fileloc_set))

fileloc = fileloc_set[0]
experiment_index = experiments_index[0]
well = 'v_'+experiment_index+'_'+videofile_set[0]
numBP_floor = 12
numBP_wall  = 7
numBP_cross = 8
numUniq = 1

hfile_floor = h5_dir+'/'+well+scorer_floor+'_el_filtered.h5'
Dataframe = pd.read_hdf(os.path.join(hfile_floor))
bodyparts = Dataframe.columns.get_level_values(2)[0:3*numBP_floor];
bodyparts_floor = (bodyparts[0::3])

uniqueID = Dataframe.columns.get_level_values(1)[0:3*numBP_floor*3];
uniqueID = (uniqueID[0::3*numBP_floor])[2]
paint = Dataframe.columns.get_level_values(2)[0:3*numBP_floor*3];
paint = paint[3*numBP_floor*2]

hfile_wall = h5_dir+'/'+well+scorer_wall+'_el_filtered.h5'
Dataframe = pd.read_hdf(os.path.join(hfile_wall))
bodyparts = Dataframe.columns.get_level_values(2)[0:3*numBP_wall];
bodyparts_wall = (bodyparts[0::3])

hfile_cross = h5_dir+'/'+well+scorer_cross+'_el_filtered.h5'
Dataframe = pd.read_hdf(os.path.join(hfile_cross))
bodyparts = Dataframe.columns.get_level_values(2)[0:3*numBP_cross];
bodyparts_cross = (bodyparts[0::3])


animalID = Dataframe.columns.get_level_values(1)[0:3*numBP_cross*3];
animalID = (animalID[0::3*numBP_cross])#[0:2]
n_ID = len(animalID)

# numBP_floor = 5
# numBP_wall  = 4
# numBP_cross = 4
# bodyparts_floor = bodyparts_floor[0:numBP_floor]
# bodyparts_wall  = bodyparts_wall[0:numBP_wall]
# bodyparts_cross = bodyparts_cross[[0,3,4,7]]

# numBP_floor = 5
# numBP_wall  = 4
# numBP_cross = 4
# bodyparts_floor = bodyparts_floor[[0,1,2,3,4]]
# bodyparts_wall  = bodyparts_wall[[0,1,2,3]]
# bodyparts_cross = bodyparts_cross[[0,3,4,7]]

# numBP_floor = 3
# numBP_wall  = 3
# numBP_cross = 4
# bodyparts_floor = bodyparts_floor[[0,3,4]]
# bodyparts_wall  = bodyparts_wall[[1,2,3]]
# bodyparts_cross = bodyparts_cross[[0,3,4,7]]

numBP_floor = 3
numBP_wall  = 3
numBP_cross = 4
bodyparts_floor = bodyparts_floor[[0,3,4]]
bodyparts_wall  = bodyparts_wall[[1,2,3]]
bodyparts_cross = bodyparts_cross[[0,3,4,7]]


# numBP_floor = 2
# numBP_wall  = 2
# numBP_cross = 4
# bodyparts_floor = bodyparts_floor[[0,4]]
# bodyparts_wall  = bodyparts_wall[[1,3]]
# bodyparts_cross = bodyparts_cross[[0,3,4,7]]


# print(bodyparts_floor1)
print(uniqueID,paint)
print(bodyparts_floor, len(bodyparts_floor))
print(bodyparts_wall, len(bodyparts_wall))
print(bodyparts_cross, len(bodyparts_cross))
# print(animalID)

num_exp = 0
for i in range(num_file):
    experiment = experiments_mscore[fileloc_set[i]]
    sexes        = experiment['sex']
    for j in range(num_well):
        sex = sexes[j]

        if sex != -1:
            num_exp = num_exp+1
            
            
total_events = np.zeros((3,num_exp)).astype(int)
genders_status   = -np.ones((num_exp)).astype(int)
genotypes_status = -np.ones((num_exp)).astype(int)
radius = -np.ones((num_exp))

ii = 0
for i in range(num_file):
    experiment  = experiments_mscore[fileloc_set[i]]
    sexes       = experiment['sex']
    genotypes   = experiment['genotype']
    events_type = experiment['events type']
    events      = experiment['events']
    circle      = experiment['arena circle']
    start_t     = np.max(experiment['starting_time'])
    well_tracker_select = np.logical_and(np.arange(frame_n)>=start_t,np.arange(frame_n)<start_t+frame_n_e)
    experiment_index = experiments_index[i]
    for j in range(num_well):

        sex = sexes[j]
        genotype = genotypes[j]

        if sex != -1:
#         if sex == 0 and genotype == 2:
            event_ts   = np.array(events[j]).astype(int)
            event_type = np.array(events_type[j]).astype(int)
            well_arena = circle[j]

            n_event = len(event_ts)
            n_event_HB  = sum(event_type==1)
            n_event_UWE = sum(event_type==2)
            n_event_L   = sum(event_type==3)
            
            n_event_HB_s  = sum(well_tracker_select[event_ts[event_type==1]])
            n_event_UWE_s = sum(well_tracker_select[event_ts[event_type==2]])
            n_event_L_s   = sum(well_tracker_select[event_ts[event_type==3]])
            
            total_events[0,ii] = n_event_HB_s
            total_events[1,ii] = n_event_UWE_s
            total_events[2,ii] = n_event_L_s

            if n_event_HB != n_event_HB_s:
                print(experiment_index+'_'+videofile_set[j],'HB mis:',n_event_HB, n_event_HB_s)
            if n_event_UWE != n_event_UWE_s:
                print(experiment_index+'_'+videofile_set[j],'UWE mis:',n_event_UWE, n_event_UWE_s)
            if n_event_L != n_event_L_s:
                print(experiment_index+'_'+videofile_set[j],'L mis:',n_event_L, n_event_L_s)

            
            genders_status[ii] = sex
            genotypes_status[ii] = genotype
            radius[ii] = (np.mean(well_arena, axis = 0))[2]
#             print(experiment_index+'_'+videofile_set[j], total_events[0,ii], total_events[1,ii], total_events[2,ii])
            ii += 1
print(ii)



# locomotion datas

np.set_printoptions(precision=2)

# window_size = np.arange(30)+1
window_size = [1,6,60]
# window_size = [1,2,60]
# window_size = [30,60]
# window_size = [1,6]
l_ws = len(window_size)

filter_status = '_filtered'
# filter_status = ''

thr_presence = 0.1
size_npz = shared_data_folder+exp_target+'_bodysize_event.npz'
loaded_size = np.load(size_npz)

d_f1f2_m_tracked, t_f1f2_m_tracked, ad_d_f1f2_m_tracked, t_ad_d_f1f2_m_tracked, t_f1f2_c_tracked, d_f1f2_tracked, t_f1f2_tracked, d_tracked, t_speed_tracked,norm_f1_f2_face_m, norm_f1_f2_chase_m, t_face_tracked, t_chase_tracked,norm_f1_f2_face_m_wocr,norm_f1_f2_chase_m_wocr, t_face_wocr_tracked, t_chase_wocr_tracked = (mcam_tools.all_tracking_data(
    h5_dir, loaded_size, filter_status, scorer_floor, scorer_wall, scorer_cross,experiments_mscore, experiments_index, fileloc_set, videofile_set, 
    animalID, bodyparts_floor, num_file,num_well, num_exp, window_size, crop_h, frame_n, frame_n_e, thr_presence))


# d_f1f2_m_tracked, t_f1f2_m_tracked, ad_d_f1f2_m_tracked, t_ad_d_f1f2_m_tracked, t_f1f2_c_tracked, d_f1f2_tracked, t_f1f2_tracked, d_tracked, t_speed_tracked,norm_f1_f2_face_m, norm_f1_f2_chase_m, t_face_tracked, t_chase_tracked,norm_f1_f2_face_m_wocr,norm_f1_f2_chase_m_wocr, t_face_wocr_tracked, t_chase_wocr_tracked = (all_tracking_data(
#     h5_dir, loaded_size, filter_status, scorer_floor, scorer_wall, scorer_cross,experiments_mscore, experiments_index, fileloc_set, videofile_set, 
#     animalID, bodyparts_floor, num_file,num_well, num_exp, window_size, crop_h, frame_n, frame_n_e, thr_presence))




np.set_printoptions(precision=3)

window_size = [1,6,60]
# window_size = [30,60]

l_ws = len(window_size)

filter_status = '_filtered'
# filter_status = ''

thr_presence = 0.1
size_npz = shared_data_folder+exp_target+'_bodysize_event.npz'
loaded_size = np.load(size_npz)
well_done_rate_f_2bp_selected, well_done_rate_w_2bp_selected, well_done_rate_c_selected, well_done_rate_temp_2bp_selected, well_done_rate_temp_selected = (mcam_tools.all_tracking_data_rate(
    h5_dir, loaded_size, filter_status, scorer_floor, scorer_wall, scorer_cross,experiments_mscore, experiments_index, fileloc_set, videofile_set, 
    animalID, bodyparts_floor, bodyparts_wall, bodyparts_cross, num_file,num_well, num_exp, window_size, crop_h, frame_n, frame_n_e, thr_presence))



print(np.mean(t_face_tracked, axis=1)/1800, np.mean(t_chase_tracked, axis=1)/1800)

# d_f1f2_m_tracked,
# t_f1f2_m_tracked,

# ad_d_f1f2_m_tracked,
# t_ad_d_f1f2_m_tracked,

# t_f1f2_c_tracked,

# d_f1f2_tracked,
# t_f1f2_tracked,

# d_tracked,
# t_speed_tracked,

# norm_f1_f2_face_m, norm_f1_f2_chase_m, t_face_tracked, t_chase_tracked,
# norm_f1_f2_face_m_wocr,norm_f1_f2_chase_m_wocr, t_face_wocr_tracked, t_chase_wocr_tracked


np.savez(shared_data_folder+exp_target+'_locomotion.npz',
         
        d_f1f2_m_tracked = d_f1f2_m_tracked, 
        t_f1f2_m_tracked = t_f1f2_m_tracked,
         
        ad_d_f1f2_m_tracked   = ad_d_f1f2_m_tracked, 
        t_ad_d_f1f2_m_tracked = t_ad_d_f1f2_m_tracked, 
         
        t_f1f2_c_tracked = t_f1f2_c_tracked,
         
        d_f1f2_tracked = d_f1f2_tracked,
        t_f1f2_tracked = t_f1f2_tracked,
        

        d_tracked = d_tracked,
        t_speed_tracked = t_speed_tracked,
        
        norm_f1_f2_chase_m = norm_f1_f2_chase_m,
        norm_f1_f2_face_m  = norm_f1_f2_face_m,
        t_chase_tracked    = t_chase_tracked,
        t_face_tracked     = t_face_tracked,
        norm_f1_f2_chase_m_wocr = norm_f1_f2_chase_m_wocr,
        norm_f1_f2_face_m_wocr  = norm_f1_f2_face_m_wocr,
        t_chase_wocr_tracked    = t_chase_wocr_tracked,
        t_face_wocr_tracked     = t_face_wocr_tracked, 

        total_events = total_events,
        genders_status = genders_status,
        genotypes_status = genotypes_status,
        radius = radius,
        )

np.savez(shared_data_folder+exp_target+'_tracking_rate.npz', 
         well_done_rate_f_2bp_selected = well_done_rate_f_2bp_selected, 
         well_done_rate_w_2bp_selected = well_done_rate_w_2bp_selected, 
         well_done_rate_c_selected = well_done_rate_c_selected, 
         well_done_rate_temp_2bp_selected = well_done_rate_temp_2bp_selected, 
         well_done_rate_temp_selected = well_done_rate_temp_selected
        )

np.set_printoptions(precision=3)
print(np.mean(well_done_rate_f_2bp_selected   ,axis=1))
print(np.mean(well_done_rate_w_2bp_selected   ,axis=1))
print(np.mean(well_done_rate_c_selected       ,axis=1))
print(np.mean(well_done_rate_temp_2bp_selected,axis=1))
print(np.mean(well_done_rate_temp_selected    ,axis=1))

t_f1f2_c_tracked_total = np.sum(t_f1f2_c_tracked,axis=2)

for i_sex in range(2):
    gender = (genders_status  ==i_sex)
    csv_block = -np.ones((np.sum(gender)+3,10*l_ws))
    csv_block[:,:] = np.NaN
    i_start = 0
    for j_gtype in range(3):
        gtype = (genotypes_status==j_gtype)
        g_g   = np.logical_and(gender, gtype)
        n_gt  = np.sum(g_g)
        j_start = 0
        for i in range(l_ws):

            
            speed = d_tracked[i,g_g]/t_speed_tracked[i,g_g]
            
#             print(i_sex,j_gtype,i, csv_block[i_start:i_start+n_gt,j_start+0].shape, speed.shape)
            
            csv_block[i_start:i_start+n_gt,j_start+0] = speed

            d_center = d_f1f2_m_tracked[i,g_g]
            csv_block[i_start:i_start+n_gt,j_start+1] = d_center

            t = t_f1f2_m_tracked[i,g_g]
            csv_block[i_start:i_start+n_gt,j_start+2] = t

            d_d_hh_tt = d_d_f1f2_hh_tt_m_tracked[i,g_g]
            csv_block[i_start:i_start+n_gt,j_start+3] = d_d_hh_tt

#             t = t_tracked_ht[i,g_g]
#             csv_block[i_start:i_start+n_gt,j_start+4] = t

            ratio = t_f1f2_c_tracked[i,g_g,1]/t_f1f2_c_tracked_total[i,g_g]
            csv_block[i_start:i_start+n_gt,j_start+4] = ratio
            
            

            
#             d_f_ht = d_f_ht_tracked[i,g_g,:]
#             csv_block[i_start:i_start+n_gt,(j_start+6):(j_start+8)] = d_f_ht
#             print(i_start, j_start)
            
            j_start += 10

#         csv_block[i_start:i_start+n_gt,5:8] = toto_event_g_g
        i_start += (sum(g_g)+1)
    np.savetxt(shared_data_folder+exp_target+'_'+sex_name[i_sex]+'_locomotion.csv', csv_block, delimiter=',', header=exp_target+'_'+sex_name[i_sex])


np.set_printoptions(precision=2)

window_size = [60]

l_ws = len(window_size)

filter_status = '_filtered'


thr_presence = 0.1
size_npz = shared_data_folder+exp_target+'_bodysize_event.npz'
loaded_size = np.load(size_npz)
d_f1f2_m_tracked, t_f1f2_m_tracked, ad_d_f1f2_m_tracked, t_ad_d_f1f2_m_tracked, t_f1f2_c_tracked, d_f1f2_tracked, t_f1f2_tracked, d_tracked, t_speed_tracked,norm_f1_f2_face_m, norm_f1_f2_chase_m, t_face_tracked, t_chase_tracked,norm_f1_f2_face_m_wocr,norm_f1_f2_chase_m_wocr, t_face_wocr_tracked, t_chase_wocr_tracked = (mcam_tools.all_tracking_data(
    h5_dir, loaded_size, filter_status, scorer_floor, scorer_wall, scorer_cross,experiments_mscore, experiments_index, fileloc_set, videofile_set, 
    animalID, bodyparts_floor, num_file,num_well, num_exp, window_size, crop_h, frame_n, frame_n_e, thr_presence))
